basic netbox 4.1 support
