+++ validity/netbox_changes/current.py
@@ -1,17 +1,9 @@
-# NetBox 4.0
+# NetBox 4.1
 from pydoc import locate as __locate
 
 from .old import *
 
 
-FieldSet = __locate("utilities.forms.rendering.FieldSet")
-plugins = __locate("netbox.plugins")
-ButtonColorChoices = __locate("netbox.choices.ButtonColorChoices")
-PluginTemplateExtension = __locate("netbox.plugins.PluginTemplateExtension")
-CF_OBJ_TYPE = "related_object_type"
-CF_CONTENT_TYPES = "object_types"
-htmx_partial = __locate("utilities.htmx.htmx_partial")
+enqueue_event = __locate("extras.events.enqueue_event")
 
-
-class BootstrapMixin:
-    pass
+QUEUE_CREATE_ACTION = "object_created"
