+++ validity/netbox_changes/oldest.py
@@ -1,16 +1,14 @@
-# NetBox 3.6
+# NetBox 3.7
 from pydoc import locate as __locate
 
 
-enqueue_object = __locate("extras.webhooks.enqueue_object")
-events_queue = __locate("netbox.context.webhooks_queue")
-EventRulesMixin = __locate("netbox.models.features.WebhooksMixin")
 BootstrapMixin = __locate("utilities.forms.BootstrapMixin")
-SavedFiltersMixin = __locate("extras.forms.mixins.SavedFiltersMixin")
 plugins = __locate("extras.plugins")
 ButtonColorChoices = __locate("utilities.choices.ButtonColorChoices")
 PluginTemplateExtension = __locate("extras.plugins.PluginTemplateExtension")
 htmx_partial = __locate("utilities.htmx.is_htmx")
+enqueue_event = __locate("extras.events.enqueue_object")
+NestedTenantSerializer = __locate("tenancy.")
 
 
 class FieldSet:
@@ -20,3 +18,4 @@ class FieldSet:
 
 CF_OBJ_TYPE = "object_type"
 CF_CONTENT_TYPES = "content_types"
+QUEUE_CREATE_ACTION = "create"
